requirements.txt
instrument-server~=1.3.7
pyserial~=3.5
